Path finding (#2)
* PathFinding

This commit introduces the PathFinding Page (Which is no working) However the boundaries are still missing.

* some performance quick fixes

* Added Barriers to the grid. Solved some Performance issues

Now it's possible to place barriers on the grid. The Algorithm will find the shortest path without going over the barriers. However if there is no path, the algorithm is going to run endlessly . Fix incoming

* Added Reset Solve Button and implemented stop when not solvable.

Now the algorithm won't run all day long when there is no solution. In this case it will reset solve. Clicking on a barrier will no delete it.

* Delete PathFindingHelp.py

No longer used. Only use case was for trying things out.

* Added presets for the Grid